Update Kiosk.java
Added user interface at the beginning as final change

Reflection:
The process so far is about cutting down on my code and making it more efficient. In the original version, I declared five integers, five Strings, and five prices. This was inefficient because I had to use the scanner class and print multiple questions asking for the quantity and names of five items. On top of asking for ten different inputs, I had to assign a price to each inputted item. This ended up causing my code to become overly repetitive and incredibly inefficient. During the time taken to debug, I thought about the repetition in my code. There was no need to type everything when I could use loops and arrays. I began by deleting my declared quantity integers and item strings. Instead, I created two arrays for quantity and item names to better organize user inputs. I utilized two “for” loops that would repeat for the length of the declared arrays. In each loop, the program would prompt the user to input how much of an item he/she wants to buy and what he/she wants to buy. My code would store the five different integer values and strings in the quantity and string arrays respectively. From there, I will multiply the stored integers at each position by the initialized prices and get a subtotal. In the printer class, I couldn’t write anything for that class on the original version due to inefficiencies. However, the current version includes a method that prints a rough outline of the receipt with the 40 char rows, a header, and quantity and price columns. In the main method, I created an object of class printer and called the print receipt method. The next steps in my thinking would be to create constructors to return the double and integer values from the Kiosk file. Then, I will create a setter method to print these values in the printer class.